drop support of Go 1.22
