+++ .github/workflows/test.yml
@@ -15,11 +15,6 @@ jobs:
           - "1.25"
           - "1.24"
           - "1.23"
-          - "1.22"
-          - "1.21"
-          - "1.20"
-          - "1.19"
-          - "1.18"
 
     steps:
       - name: Check out code into the Go module directory
