+++ go.mod
@@ -1,3 +1,3 @@
 module github.com/shogo82148/go-retry/v2
 
-go 1.18
+go 1.23
